+++ app/settings/@modal/(..)logout/page.tsx
@@ -1,4 +1,4 @@
-import Logout from "@/components/Logout";
+import Logout from "@/components/Setting/Logout";
 import Modal from "@/components/@common/Modal";
 
 const Page = () => {
