+++ components/Setting/Logout/index.tsx
