+++ components/Setting/MyProfile/index.tsx
@@ -1,4 +1,4 @@
-import * as styles from "@/components/MyProfile/style.css";
+import * as styles from "@/components/Setting/MyProfile/style.css";
 import data from "./mockData.json";
 
 const MyProfile = () => {
