+++ components/Setting/MypetCarousel/Card/style.css.ts
